Fix issues with sign conversion

+++ src/libnav/libnav/navaid_db.hpp
@@ -184,7 +184,7 @@ namespace libnav
 
 		// get_wpt_data returns 0 if waypoint is not in the database. 
 		// Otherwise, returns number of items written to out.
-		int get_wpt_data(std::string& id, std::vector<waypoint_entry_t>* out, 
+		size_t get_wpt_data(std::string& id, std::vector<waypoint_entry_t>* out, 
 			std::string area_code="", std::string country_code="", 
 			NavaidType type=NavaidType::NAVAID, 
 			navaid_filter_t filt_func=default_navaid_filter, void* ref=NULL);
@@ -198,7 +198,7 @@ namespace libnav
 			@return: number of items written to out.
 		*/
 
-		int get_wpt_by_awy_str(std::string& awy_str, std::vector<waypoint_entry_t>* out);
+		size_t get_wpt_by_awy_str(std::string& awy_str, std::vector<waypoint_entry_t>* out);
 
 		/*
 			Function: get_wpt_by_hold_str
@@ -209,7 +209,7 @@ namespace libnav
 			@return: number of items written to out.
 		*/
 
-		int get_wpt_by_hold_str(std::string& hold_str, std::vector<waypoint_entry_t>* out);
+		size_t get_wpt_by_hold_str(std::string& hold_str, std::vector<waypoint_entry_t>* out);
 
 		std::string get_fix_desc(waypoint_t& fix);
